requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Matrix SDK
matrix-nio==0.21.0
aiofiles==23.2.1

# Database
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.12.1

# LiveKit
livekit==0.10.0
livekit-api==0.9.0

# HTTP client
httpx==0.25.2

# Logging
structlog==23.2.0



